Created Dashing and Invulnerability Functionality
The iFrames work, but for the dash I'm trying to figure out how to make it so that it goes in a certain direction and you can do it after spawning.

I'm pretty sure my code is also supposed to change the opacity but for now it changes the players color. :(